test for existence of requirements.txt/setup.py

+++ vdist/templates/debian.sh
@@ -20,7 +20,13 @@ virtualenv .
 
 source bin/activate
 
-pip install -r requirements.txt
+if [ -f "requirements.txt" ]; then
+    pip install -r requirements.txt
+fi
+
+if [ -f "setup.py" ]; then
+    python setup.py install
+fi
 
 cd ..
